Start moving model/view/urls
