+++ settings.py
@@ -225,6 +225,7 @@ HQ_APPS = (
     'corehq.apps.hqmedia',
     'corehq.apps.loadtestendpoints',
     'corehq.apps.locations',
+    'corehq.apps.products',
     'corehq.apps.commtrack',
     'corehq.apps.consumption',
     'couchforms',
@@ -958,6 +959,7 @@ COUCHDB_APPS = [
     'phone',
     'pillowtop',
     'pillow_retry',
+    'products',
     'reminders',
     'reports',
     'sofabed',
